third_party/blink/web_tests/external/wpt/trusted-types/trusted-types-reporting-for-Element-setAttribute-expected.txt

This is a testharness.js-based test.
[FAIL] Violation report for HTMLIFrameElement.setAttribute('srcdoc', plain_string)
  assert_equals: expected "HTMLIFrameElement srcdoc|AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA" but got "Element setAttribute|AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA"
[FAIL] Violation report for HTMLIFrameElement.setAttributeNS(null, 'srcdoc', plain_string)
  assert_equals: expected "HTMLIFrameElement srcdoc|AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA" but got "Element setAttributeNS|AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA"
[FAIL] Violation report for Element.setAttribute('onclick', plain_string)
  assert_equals: expected "Element onclick|AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA" but got "Element setAttribute|AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA"
[FAIL] Violation report for Element.setAttributeNS(null, 'onclick', plain_string)
  assert_equals: expected "Element onclick|AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA" but got "Element setAttributeNS|AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA"
[FAIL] Violation report for HTMLScriptElement.setAttribute('src', plain_string)
  assert_equals: expected "HTMLScriptElement src|AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA" but got "Element setAttribute|AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA"
[FAIL] Violation report for HTMLScriptElement.setAttributeNS(null, 'src', plain_string)
  assert_equals: expected "HTMLScriptElement src|AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA" but got "Element setAttributeNS|AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA"
[FAIL] Violation report for SVGScriptElement.setAttribute('href', plain_string)
  assert_equals: expected "SVGScriptElement href|AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA" but got "Element setAttribute|AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA"
[FAIL] Violation report for SVGScriptElement.setAttributeNS(null, 'href', plain_string)
  assert_equals: expected "SVGScriptElement href|AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA" but got "Element setAttributeNS|AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA"
[FAIL] Violation report for SVGScriptElement.setAttributeNS(http://www.w3.org/1999/xlink, 'href', plain_string)
  assert_equals: expected "SVGScriptElement href|AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA" but got "Element setAttributeNS|AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA"
Harness: the test ran to completion.
